db refactor

+++ rustytime/src/handlers/github.rs
@@ -129,7 +129,7 @@ pub async fn callback(
     // create or update user in database
     let user = match User::create_or_update(
         &mut conn,
-        user_info.id as i32,
+        user_info.id as i64,
         &user_info.login,
         &user_info.avatar_url,
     ) {
